+++ grid2op/Agent/TopologyGreedy.py
@@ -23,14 +23,14 @@ class TopologyGreedy(GreedyAgent):
     """
     def __init__(self, action_space):
         GreedyAgent.__init__(self, action_space)
-        self.li_actions = None
+        self.tested_action = None
 
     def _get_tested_action(self, observation):
-        if self.li_actions is None:
+        if self.tested_action is None:
             res = [self.action_space({})]  # add the do nothing
             # better use "get_all_unitary_topologies_set" and not "get_all_unitary_topologies_change"
             # maybe "change" are still "bugged" (in the sens they don't count all topologies exactly once)
             res += self.action_space.get_all_unitary_topologies_set(self.action_space)
-            self.li_actions = res
-        return self.li_actions
+            self.tested_action = res
+        return self.tested_action